+++ pyscript/core/position.py
@@ -25,75 +25,75 @@ class PysPosition(Pys):
     def __repr__(self):
         return f'<Position({self.start!r}, {self.end!r}) from {self.file.name!r}>'
 
-    def format_arrow(self, colored=True):
-        if self.is_positionless:
-            return ''
+def format_arrow(position, colored=True):
+    if position.is_positionless:
+        return ''
 
-        if colored:
-            reset = acolor('reset')
-            bred = acolor('red', style=BOLD)
-        else:
-            reset = ''
-            bred = ''
+    if colored:
+        reset = acolor('reset')
+        bred = acolor('red', style=BOLD)
+    else:
+        reset = ''
+        bred = ''
 
-        line_start = self.start_line
-        line_end = self.end_line
-        column_start = self.start_column
-        column_end = self.end_column
+    line_start = position.start_line
+    line_end = position.end_line
+    column_start = position.start_column
+    column_end = position.end_column
 
-        text = self.file.text
+    text = position.file.text
 
-        start = text.rfind('\n', 0, self.start) + 1
+    start = text.rfind('\n', 0, position.start) + 1
+    end = text.find('\n', start + 1)
+    if end == -1:
+        end = len(text)
+
+    if text[position.start:position.end] in ('', '\n'):
+        if position.start > start:
+            line = text[start:end].lstrip().replace('\t', ' ')
+            return f'{line}\n{bred}{" " * len(line)}^{reset}'
+        return f'\n{bred}^{reset}'
+
+    result = []
+    lines = []
+    count = line_end - line_start + 1
+
+    for i in range(count):
+        line = text[start:end].lstrip('\n')
+
+        lines.append(
+            (
+                line,
+                len(line.lstrip()),
+                column_start - 1 if i == 0 else 0,
+                column_end - 1 if i == count - 1 else len(line)
+            )
+        )
+
+        start = end
         end = text.find('\n', start + 1)
         if end == -1:
             end = len(text)
 
-        if text[self.start:self.end] in ('', '\n'):
-            if self.start > start:
-                line = text[start:end].lstrip().replace('\t', ' ')
-                return f'{line}\n{bred}{" " * len(line)}^{reset}'
-            return f'\n{bred}^{reset}'
+    minimum_indent = min(len(line) - line_code_length for line, line_code_length, _, _ in lines)
 
-        result = []
-        lines = []
-        count = line_end - line_start + 1
+    for i, (line, line_code_length, start, end) in enumerate(lines):
+        line = line[minimum_indent:]
+        er = end - minimum_indent
 
-        for i in range(count):
-            line = text[start:end].lstrip('\n')
+        if i == 0:
+            sr = start - minimum_indent
 
-            lines.append(
-                (
-                    line,
-                    len(line.lstrip()),
-                    column_start - 1 if i == 0 else 0,
-                    column_end - 1 if i == count - 1 else len(line)
-                )
-            )
+            arrow = '^' * (end - start)
+            line = f'{line[:sr]}{bred}{line[sr:er]}{reset}{line[er:]}\n{" " * sr}{bred}{arrow}{reset}'
 
-            start = end
-            end = text.find('\n', start + 1)
-            if end == -1:
-                end = len(text)
+        else:
+            indent = len(line) - line_code_length
 
-        minimum_indent = min(len(line) - line_code_length for line, line_code_length, _, _ in lines)
+            arrow = '^' * (end - start - (minimum_indent + indent))
+            line = f'{line[:indent]}{bred}{line[indent:er]}{reset}{line[er:]}\n{" " * indent}{bred}{arrow}{reset}'
 
-        for i, (line, line_code_length, start, end) in enumerate(lines):
-            line = line[minimum_indent:]
-            er = end - minimum_indent
+        if arrow:
+            result.append(line)
 
-            if i == 0:
-                sr = start - minimum_indent
-
-                arrow = '^' * (end - start)
-                line = f'{line[:sr]}{bred}{line[sr:er]}{reset}{line[er:]}\n{" " * sr}{bred}{arrow}{reset}'
-
-            else:
-                indent = len(line) - line_code_length
-
-                arrow = '^' * (end - start - (minimum_indent + indent))
-                line = f'{line[:indent]}{bred}{line[indent:er]}{reset}{line[er:]}\n{" " * indent}{bred}{arrow}{reset}'
-
-            if arrow:
-                result.append(line)
-
-        return '\n'.join(result).replace('\t', ' ')
+    return '\n'.join(result).replace('\t', ' ')